requirements.txt
# Core ML/DL libraries (optimized versions)
torch>=2.1.0  # PyTorch 2.1+ for torch.compile and better performance
torchaudio>=2.1.0
pytorch-lightning>=2.1.0
transformers>=4.35.0

# Training acceleration libraries
deepspeed>=0.11.0  # For ZeRO optimizer and mixed precision
accelerate>=0.24.0  # HuggingFace acceleration
flash-attn>=2.3.0  # Flash Attention 2 for faster attention
apex  # NVIDIA Apex for mixed precision (optional, if available)
bitsandbytes>=0.41.0  # 8-bit optimizers for memory efficiency

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
datasets>=2.14.0
huggingface-hub>=0.16.4

# Audio processing for speech recognition
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0
resampy>=0.4.0
pyannote.audio>=3.0.0
speechbrain>=0.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# Optimization & Deployment
tensorrt>=8.6.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
triton-client[all]>=2.35.0

# Time series specific
pytorch-forecasting>=1.0.0
sktime>=0.24.0
tsfresh>=0.20.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0
typer>=0.9.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# GPU monitoring
pynvml>=11.5.0
gpustat>=1.1.0